35 students refactoring

+++ src/main/java/com/schoolplus/office/web/models/AnnouncementImageDto.java
@@ -15,6 +15,6 @@ public class AnnouncementImageDto {
 
     private String imageUrl;
 
-    private Integer imageSize;
+    private Long imageSize;
 
 }
